Decoupled the DriverInterface from the NodeElement
This avoids a circular dependency. The driver interface now returns an
array of XPath for found elements, and Element::find() instantiates the
NodeElement for them.
The driver is now the low level API which does not depend on any other
part of Mink (but used by other parts).
The ElementFinder is responsible for instantiating the NodeElement.