Reverted the cloning implementation for the Session
Cloning a Session would require making sure that all drivers all
cloneable, which would then require that all dependencies of these drivers
maintaining a state are cloneable as well. But this is not the case in the
existing drivers.
Thus, cloning the session would require checking that the object graph is
cloned in a consistent way (i.e. all references to the driver are
referencing a single clone for instance). The changes started for Mink 2.0
to have a cleaner architecture are making it really hard to ensure this.
So the decision was to revert this change before creating a stable release
including it.
Reverts #368
Closes #445