Split the MouseOver event into several tests
Selenium1 and Selenium2 drivers are both marking this test as incomplete
and duplicating part of it for the events they support, so that these
events can be tested without being skipped because of other unsupported
events. Because of this, it does not make sense to have them in the same
test.