Added tests for invalid usages of the drivers
These tests cover calling the driver for XPath not matching elements on
the page (case which should not happen in normal usage of Mink given that
the driver is accessed from found elements most of the time) or for XPath
matching an invalid element (trying to check a div for instance).
These tests ensure that a DriverException is thrown rather than doing an
undefined behavior (potentially a fatal error).