Removed the session uniqueId
This uniqueId is not used anywhere in Mink, and the value is not even
guaranteed to be unique. On windows, uniqid does not have enough entropy
to ensure that the testsuite passes consistently for instance.
If the external code needs a way to identify a Session instance uniquely
as a scalar, it is better to use spl_object_hash, which will not change as
long as the Session is not garbage collected (at which point it is
probably the end of the test process).