Improved the description of what resetting a driver means
This documents that only a few actions are allowed on a driver which has
been started or reset until a page is visited. This matches the behavior
of browsers, where you cannot perform actions on a page before opening
one.
A new test is added to ensure that the basic auth is properly reset when
resetting the driver.