Changed the architecture of the driver testsuite
Instead of having all driver tests in a single abstract testcase becoming
so huge it is hard to maintain, the driver testsuite is now composed by a
set of concrete testcases and a config object to build the driver and
adapt the testsuite.
This makes the maintenance of the driver testsuite easier by having
smaller targetted testcases.
It also separates the driver testsuite from the Mink testsuite (required
now that testcases are not abstract anymore in the driver testsuite),
which will avoid confusion for contributors.