Implement the conversion of XPath to NodeElement in CoreDriver
The goal is to ease the development of Mink 2.0 by allowing drivers to
be compatible with both Mink 1.7 and 2.0 on a single code base. The only
change required for drivers is now handled at the level of the
CoreDriver shipped in Mink when drivers should to override
findElementXpaths rather than find.