Allow selectors and replacements to be registered late
This allows projects using Mink to create their Named and Partial
selectors, and then to add additional replacements and selectors to be
registered for subsequent use.

This is particularly useful in projects with a distributed plugin
architecture which make use of replacements and selectors without
abusing Reflection to do so.